human/requirements.txt
# Core Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchmetrics>=0.11.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.7.0
h5py>=3.6.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
Pillow>=8.3.0

# Model Interpretability
shap>=0.41.0

# Utilities
tqdm>=4.62.0
jupyter>=1.0.0
ipython>=7.0.0
PyYAML>=5.4.0

# Optional: For GPU support
# Remove comment if using CUDA
# torch-cuda-extension>=0.2.0